Modularize E2ESaveYourPasswordStackNavigator
